requirements.txt
streamlit
pillow
matplotlib
scikit-learn